+++ server/CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning][semver].
 ### Added:
 
 - switch from `requests` to `aiohttp`. Doing http requests asynchronously and in the case of pypi diagnostics, concurrently.
+- diagnostic: lower severity of option redefined to same value to information level
+- diagnostic: add hint level diagnostic for every redefined value
 
 ### Fixed: