+++ server/CHANGELOG.md
@@ -13,6 +13,7 @@ and this project adheres to [Semantic Versioning][semver].
 - diagnostic: lower severity of option redefined to same value to information level
 - diagnostic: add hint level diagnostic for every redefined value
 - cli: new `--log-pygls` flag to include pygls logs. By default they are not included.
+- diagnostic: report unknown pypi projects or versions
 
 ### Fixed: