+++ server/CHANGELOG.md
@@ -19,6 +19,7 @@ and this project adheres to [Semantic Versioning][semver].
 
 - pypi: fix a crash when unknown project is listed in `versions`
 - diagnostic: don't emit false positives for profiles generated with jinja
+- fix crash with empty `${buildout:extends}`
 
 ## [0.7.1] - 2022-05-08