+++ server/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning][semver].
 
 ## [Unreleased]
 
+## [0.8.0] - 2022-07-02
+
 ### Added:
 
 - switch from `requests` to `aiohttp`. Doing http requests asynchronously and in the case of pypi diagnostics, concurrently.
@@ -153,4 +155,5 @@ and this project adheres to [Semantic Versioning][semver].
 [0.6.2]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.6.1...v0.6.2
 [0.7.0]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.6.2...v0.7.0
 [0.7.1]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.7.0...v0.7.1
-[unreleased]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.7.1...master
+[0.8.8]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.7.1...v0.8.0
+[unreleased]: https://github.com/perrinjerome/vscode-zc-buildout/compare/v0.8.0...main